Optional local volume for pravega up.

+++ pravega-docker/up.sh
@@ -14,9 +14,14 @@ set -ex
 ROOT_DIR=$(readlink -f $(dirname $0))
 cd ${ROOT_DIR}
 export HOST_IP=${HOST_IP:-192.168.120.128}
-export PRAVEGA_LTS_PATH=${PRAVEGA_LTS_PATH:-/tmp/pravega-lts}
-docker-compose down
-sudo rm -rf ${PRAVEGA_LTS_PATH}
+export PRAVEGA_LTS_PATH=${PRAVEGA_LTS_PATH-/tmp/pravega-lts}
+docker-compose down -v
+
+if ! [ -z "${PRAVEGA_LTS_PATH}" ]
+then
+    sudo rm -rf ${PRAVEGA_LTS_PATH}
+fi
+
 docker-compose up -d
 sleep 20s
 curl -X POST -H "Content-Type: application/json" -d '{"scopeName":"examples"}' http://localhost:10080/v1/scopes
